Add files via upload
I have been looking over these files in order to figure out more how the touch screen works.
From what I can tell, the buttons have to be created one by one, and require codes for the colors. I have been playing around with it in order to create set symbols and buttons that we can randomize in the future. (random using a function, not a floating ground). 
It seems to be pretty simple, however it takes quite a while to get something that looks good.