Add files via upload
The keypad 2.0 sketch is the code for the keypad that sends information over a two wire serial connection. it was originally the phonecal file, however this is modified for serial, as well as the send, del, and clr, functions. 

README.txt is the "users manual" for our device. it contains instructions.